// src/app/models/Usuario.model.ts
export interface Usuario {
    id?: number;
    username?: string;
    password?: string;
    firstname?: string;
    lastname?: string;
    direccion?: string;
    telefono?: string;
    usuario_creacion?: string;
    fecha_creacion?: Date;
    usuario_modificacion?: string;
    fecha_modificacion?: Date;
    foto?: string;
    status?: string;
  }
  